requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1

# LLM API
groq==0.4.1

# Web interface
streamlit==1.29.0

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Data processing
numpy==1.24.3
pandas==2.0.3

# Optional: for better PDF handling
pdfplumber==0.10.3
